refactor: unify config loader module
