+++ tests/live_price_ticker_simulation.py
@@ -8,7 +8,7 @@ from alert_core.alert_repository import AlertRepository
 from alert_core.alert_enrichment_service import AlertEnrichmentService
 from xcom.notification_service import NotificationService
 from data.alert import Alert, AlertType
-from utils.config_loader import load_config
+from config.config_loader import load_config
 from core.core_imports import log, retry_on_locked
 
 # Mock DataLocker with Live Price Updates
